drivers:rf-transceiver:ad9361:iio_ad9361.c: Change type to fix warnings
ad9361_get_trx_clock_chain() expects 'uint32_t' not 'unsigned long'.
This fixes several warnings of type:

warning: passing argument 3 of ‘ad9361_get_trx_clock_chain’ from incompatible pointer type

and

warning: format ‘%u’ expects argument of type ‘unsigned int’, but argument 4 has type ‘long unsigned int’

+++ drivers/rf-transceiver/ad9361/iio_ad9361.c
@@ -1360,7 +1360,7 @@ static int get_rx_path_rates(void *device, char *buf, uint32_t len,
 			     const struct iio_ch_info *channel, intptr_t priv)
 {
 	struct ad9361_rf_phy *ad9361_phy = (struct ad9361_rf_phy *)device;
-	unsigned long clk[6];
+	uint32_t clk[6];
 	int ret = ad9361_get_trx_clock_chain(ad9361_phy, clk, NULL);
 
 	if (ret < 0)
@@ -1557,7 +1557,7 @@ static int get_tx_path_rates(void *device, char *buf, uint32_t len,
 			     const struct iio_ch_info *channel, intptr_t priv)
 {
 	struct ad9361_rf_phy *ad9361_phy = (struct ad9361_rf_phy *)device;
-	unsigned long clk[6];
+	uint32_t clk[6];
 	int ret = ad9361_get_trx_clock_chain(ad9361_phy, NULL, clk);
 
 	if (ret < 0)
